requirements.txt
openai>=1.91.0
Pillow>=11.2.1
pytesseract>=0.3.13
python-dotenv>=1.1.1
reportlab>=4.4.2
requests>=2.32.4
pytz>=2024.1
tqdm>=4.67.1 